Extracting Proxy config into a separate object for a better readability of the code
